+++ android-dex-ui/src/android_dex_ui/__init__.py
@@ -1,3 +1,3 @@
 """Android-DEX UI."""
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"